requirements.txt
arrow
asgiref
beautifulsoup4
binaryornot
certifi
chardet
charset-normalizer
click
colorama
cookiecutter
crispy-bootstrap5
crispy-tailwind
distlib
dj-database-url
Django
django-appconf
django-bootstrap-v5
django-browser-reload
django-compressor
django-cors-headers
django-crispy-forms
django-environ
django-filter
django-tailwind
django-tailwindcss
djangorestframework
djangorestframework-simplejwt
filelock
idna
Jinja2
Markdown
markdown-it-py
MarkupSafe
mdurl
MouseInfo
multipledispatch
mysql-connector-python
numpy
opencv-contrib-python
pillow
pipenv
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
Pygments
PyJWT
PyMsgBox
pyperclip
PyRect
PyScreeze
python-dateutil
python-slugify
pytweening
PyYAML
rcssmin
requests
rich
rjsmin
setuptools
six
sortedcontainers
soupsieve
sqlparse
sslcommerz-lib
text-unidecode
types-python-dateutil
typing_extensions
tzdata
urllib3
virtualenv
wheel